requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
django-cors-headers
Pillow
